New 'note' content type with sequential browsing
This allows us to create courses with multiple squential notes that have rich text content.

Note: as part of this, I've moved the responsibility for rich text processing – of the rich text fields from Storyblok – to the frontend, instead of the backend.

This commit also adds in "back" navigational aids.