Backend/API to support "preview mode" of content (resources, themes and pages).
IMPORTANT: now only **published** content will be returned by default, unless preview mode is "activated" by the frontend.

Two key aspects control _how_ and _who_ gets to see **draft** content from Storyblok:

1. To signal that an API request should work in preview mode and attempt to fetch **draft** content, the client needs to set a new header in each API request: `X-PREVIEW-MODE` (value can be anything apart from a falsey value or the string 'false').

2. Only authenticated API requests (i.e. with the Firebase Auth token) will be allowed to do this, AND only certain users that have their email address present in the new env var `CONTENT_EDITOR_EMAILS` (which the backend safelists against).

So if the preview header is set and, either no auth is provided or auth for a user that's not allowed is provided, then a `403 Forbidden` is returned.